Refactor the db types a bit:
* Split db types files (for easier management)
* Update base database and inventory/orders databaseses (move some types around)
* Update e2e package with appropriate types/methods (InventoryDatabaseInterface)
* Update web-client package with appropriate types/methods (InventoryDatabaseInterface)
