+++ apps/web-client/src/lib/db/index.ts
@@ -1,10 +1,10 @@
 import pouchdb from "pouchdb";
 
-import { newInventoryDatabaseInterface, type DatabaseInterface } from "@librocco/db";
+import { newInventoryDatabaseInterface, type InventoryDatabaseInterface } from "@librocco/db";
 
 import { LOCAL_POUCH_DB_NAME } from "$lib/constants";
 
-let db: DatabaseInterface | undefined = undefined;
+let db: InventoryDatabaseInterface | undefined = undefined;
 
 /**
  * We're using createDB() instead of exporting the db itself because we don't want to
@@ -14,7 +14,7 @@ let db: DatabaseInterface | undefined = undefined;
  * It should be initialized in the browser environment and is idempotent (if the db is already instantiated, it will return the existing instance).
  * This is to prevent expensive `db.init()` operations on each route change.
  */
-export const createDB = async (): Promise<DatabaseInterface> => {
+export const createDB = async (): Promise<InventoryDatabaseInterface> => {
 	if (db) {
 		return db;
 	}
@@ -30,7 +30,7 @@ export const createDB = async (): Promise<DatabaseInterface> => {
  * Get db returns the instantiated db instance. It's safe to run this in any environment (browser/server)
  * as, it will simply return undefined if the db is not instantiated.
  */
-export const getDB = (): DatabaseInterface | undefined => {
+export const getDB = (): InventoryDatabaseInterface | undefined => {
 	return db;
 };
