+++ pkg/db/src/__tests__/tests.ts
@@ -3,11 +3,11 @@ import { expect } from "vitest";
 import { BehaviorSubject, firstValueFrom, switchMap } from "rxjs";
 import { Search } from "js-search";
 
-import { NoteState, testUtils } from "@librocco/shared";
+import { NoteState, testUtils, VolumeStock } from "@librocco/shared";
 
 import { DocType } from "../enums";
 
-import { BookEntry, InNoteMap, NavMap, VersionedString, VolumeStock, VolumeStockClient, WarehouseData } from "@/types";
+import { BookEntry, InNoteMap, NavMap, VersionedString, VolumeStockClient, WarehouseData } from "@/types";
 import { TestFunction } from "@/test-runner/types";
 
 /** TODO: This will only work for "v1" at this point, make this more versatile */
