+++ pkg/db/src/types/orders.ts
@@ -1,0 +1,7 @@
+import { DatabaseInterface as BaseDatabaseInterface } from "./misc";
+
+export type OrdersDatabaseInterface = BaseDatabaseInterface;
+
+export interface NewOrdersDatabase {
+	(db: PouchDB.Database): OrdersDatabaseInterface;
+}
